+++ src/main/java/roomescape/waiting/controller/dto/request/WaitingCreateRequest.java
@@ -1,4 +1,4 @@
-package roomescape.waiting.service.dto.request;
+package roomescape.waiting.controller.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
